feat: rename classes to backends
